Fix wrong argument fetch in buffer printf routine
* libwget/buffer_printf.c (wget_buffer_vprintf_append): Fetch long value
  from stack instead of long long with %l[dioux]

Found on Solaris 11 SPARC

+++ libwget/buffer_printf.c
@@ -362,7 +362,7 @@ size_t wget_buffer_vprintf_append(wget_buffer_t *buf, const char *fmt, va_list a
 				argu = (unsigned long long)arg;
 			} else {
 				p++;
-				arg = (long)va_arg(args, long long);
+				arg = (long)va_arg(args, long);
 				argu = (unsigned long)arg;
 			}
 			break;
